+++ packages/backend/src/models/entities/drive-file.ts
@@ -1,7 +1,7 @@
 import { PrimaryColumn, Entity, Index, JoinColumn, Column, ManyToOne } from 'typeorm';
+import { id } from '../id.js';
 import { User } from './user.js';
 import { DriveFolder } from './drive-folder.js';
-import { id } from '../id.js';
 
 @Entity()
 @Index(['userId', 'folderId', 'id'])
@@ -155,6 +155,18 @@ export class DriveFile {
 		comment: 'Whether the DriveFile is NSFW.',
 	})
 	public isSensitive: boolean;
+
+	@Column('boolean', {
+		default: false,
+	})
+	public forceIsSensitive: boolean;
+
+	@Index()
+	@Column('boolean', {
+		default: false,
+		comment: 'Whether the DriveFile is NSFW. (predict)',
+	})
+	public predictedIsSensitive: boolean;
 
 	/**
 	 * 外部の(信頼されていない)URLへの直リンクか否か